i changed the data types of the models so a much more approprtiate data type that stores the amounts of the hoursworked hourly rate in order to return a totalamount as a decimal
also made adjustments to the claims dashboard to show the total amount that the user has claimed

also added appropriate pop-up messeges

+++ Models/Calculations.cs
@@ -8,14 +8,14 @@ namespace Part1ex.Models
         [Key]
        public int claimid { get; set; }
         [Required]
-        public int HoursWorked { get; set; }
+        public decimal HoursWorked { get; set; }
 
 
         [Required]
-        public int HourlyRate { get; set; }
+        public decimal HourlyRate { get; set; }
 
 
-        public int TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
 
         //expended to include the datetime documents claimstatus
         public DateTime ClaimDate { get; set; } = DateTime.Now;
